adding function update package for admin

+++ src/main/java/com/restaurent/manager/dto/request/restaurant/RestaurantUpdateRequest.java
@@ -21,7 +21,7 @@ public class RestaurantUpdateRequest {
     private String province;
     private String district;
     @NotNull(message = "id pack is require to update")
-    private int packId;
+    private Long packId;
     @NotNull(message = "months is require for extend package")
     @Min(value = 1, message = "extend package months must be equal or greater than 1")
     private int months;
